+++ presto/testing/common/schemas/tpch/lineitem.sql
@@ -1,8 +1,8 @@
-CREATE TABLE hive.{schema}.lineitem     (
+CREATE TABLE hive.{schema}.lineitem (
     l_orderkey BIGINT NOT NULL,
     l_partkey BIGINT NOT NULL,
     l_suppkey BIGINT NOT NULL,
-    l_linenumber BIGINT NOT NULL,
+    l_linenumber INTEGER NOT NULL,
     l_quantity DOUBLE NOT NULL,
     l_extendedprice DOUBLE NOT NULL,
     l_discount DOUBLE NOT NULL,
@@ -15,4 +15,4 @@ CREATE TABLE hive.{schema}.lineitem     (
     l_shipinstruct VARCHAR NOT NULL,
     l_shipmode VARCHAR NOT NULL,
     l_comment VARCHAR NOT NULL
-)     WITH (FORMAT = 'PARQUET', EXTERNAL_LOCATION = 'file:{file_path}')
+) WITH (FORMAT = 'PARQUET', EXTERNAL_LOCATION = 'file:{file_path}')
