+++ presto/testing/common/schemas/tpch/nation.sql
@@ -1,6 +1,6 @@
-CREATE TABLE hive.{schema}.nation     (
-    n_nationkey INTEGER NOT NULL,
+CREATE TABLE hive.{schema}.nation (
+    n_nationkey BIGINT NOT NULL,
     n_name VARCHAR NOT NULL,
-    n_regionkey INTEGER NOT NULL,
+    n_regionkey BIGINT NOT NULL,
     n_comment VARCHAR NOT NULL
-)     WITH (FORMAT = 'PARQUET', EXTERNAL_LOCATION = 'file:{file_path}')
+) WITH (FORMAT = 'PARQUET', EXTERNAL_LOCATION = 'file:{file_path}')
