+++ presto/testing/common/schemas/tpch/region.sql
@@ -1,5 +1,5 @@
-CREATE TABLE hive.{schema}.region     (
-    r_regionkey INTEGER NOT NULL,
+CREATE TABLE hive.{schema}.region (
+    r_regionkey BIGINT NOT NULL,
     r_name VARCHAR NOT NULL,
     r_comment VARCHAR NOT NULL
-)     WITH (FORMAT = 'PARQUET', EXTERNAL_LOCATION = 'file:{file_path}')
+) WITH (FORMAT = 'PARQUET', EXTERNAL_LOCATION = 'file:{file_path}')
